requirements.txt
django==3.2
Pillow~=10.0
environs[django]~=9.3
more_itertools~=10.1.0
django-phonenumber-field~=7.1.0